Now redirects in Spring-Way
