+++ bangladesh-green/build.gradle
@@ -29,8 +29,12 @@ dependencies {
     compile 'org.webjars:bootstrap:3.3.5'
     compile 'org.webjars:jquery:2.1.4'
     compile 'io.jsonwebtoken:jjwt:0.6.0'
+    compile 'org.springframework.boot:spring-boot-starter-web:1.2.6.RELEASE'
     compile "org.springframework.boot:spring-boot-starter-data-mongodb"
     compile 'de.flapdoodle.embed:de.flapdoodle.embed.mongo:1.50.0'
+    compile 'jstl:jstl:1.2'
+    compile 'org.apache.tomcat.embed:tomcat-embed-jasper:8.0.28'
+
