+++ bangladesh-green/src/main/java/urlshortener/bangladeshgreen/web/JSPController.java
@@ -1,0 +1,11 @@
+package urlshortener.bangladeshgreen.web;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.ModelAndView;
+
+public class JSPController {
+
+
+    
+}
