+++ bangladesh-green/src/main/resources/application.properties
@@ -1,0 +1,29 @@
+# In this file, we can find the different configurable parameters of the application
+
+# View Resolver
+# This is the configuration related with the Web Model View Controller (MVC) Properties. We specify that
+# the Spring's dispatcher will resolver the logical view names to physical view pages
+# adding "/WEB-INF/jsp" before the names (spring.view.prefix) and ".jsp" after them
+
+spring.view.prefix: /WEB-INF/jsp/
+spring.view.suffix: .jsp
+
+spring.mvc.view.prefix=/WEB-INF/jsp/
+spring.mvc.view.suffix=.jsp
+
+
+
+# Debug level
+# The available logging levels are: "TRACE, DEBUG, INFO, WARN, ERROR, FATAL, OFF". The level
+# defines the minimum set of levels recognized by the system.
+# The level "INFO" shows the general application flow; information which should help any observer
+# understand what the application is doing in general.
+# By default, we will see the logs in the console, they will not be written in a file.
+logging.level.org.springframework.web: INFO
+#File where the log info will be written
+logging.file: info_app.log
+
+# App messages
+# This app message is displayed by default when the app is running. The user must to go to 'localhost:8080' to
+# see it.
+app.message: Hola alumnos
