+++ bangladesh-green/src/main/webapp/WEB-INF/.gitignore
@@ -1,0 +1,1 @@
+/web.xml
