+++ bangladesh-green/src/test/java/urlshortener/bangladeshgreen/web/UrlShortenerControllerTest.java
@@ -59,89 +59,7 @@ public class UrlShortenerControllerTest {
 		this.mockMvc = MockMvcBuilders.standaloneSetup(urlShortener).build();
 	}
 
-	@Test
-	/*
-	Test that REDIRECT over a NON-PRIVATE link redirects if KEY EXISTS.
-	 */
-	public void thatRedirectToReturnsTemporaryRedirectIfKeyExists()
-			throws Exception {
 
-		//Mock URLrepository response to someUrl.
-		when(shortURLRepository.findByHash("someKey")).thenReturn(someUrl());
-
-		//Test redirection
-		mockMvc.perform(get("/{id}", "someKey")).andDo(print())
-				.andExpect(status().isTemporaryRedirect())
-				.andExpect(redirectedUrl("http://www.google.es"));
-	}
-
-	@Test
-	/*
-	Test that REDIRECT over a PRIVATE link redirects if KEY EXISTS
-	and Private Token IS CORRECT.
-	 */
-	public void thatRedirectToPrivateReturnsTemporaryRedirectIfKeyExistsAndPrivateTokenCorrect()
-			throws Exception {
-
-		//Mock URLrepository response to someUrl.
-		when(shortURLRepository.findByHash("someKey")).thenReturn(somePrivateUrl());
-
-		//Test redirection
-		mockMvc.perform(get("/{id}", "someKey")
-				.param("privateToken","privateToken"))
-				.andDo(print())
-				.andExpect(status().isTemporaryRedirect())
-				.andExpect(redirectedUrl("http://www.google.es"));
-	}
-
-	@Test
-	/*
-	Test that REDIRECT over a PRIVATE link gives error 401 if key exists
-	and Private Token IS NOT CORRECT.
-	 */
-	public void thatRedirectToPrivateReturnsTemporaryRedirectIfKeyExistsAndPrivateTokenIncorrect()
-			throws Exception {
-
-		//Mock URLrepository response to a private URL.
-		when(shortURLRepository.findByHash("someKey")).thenReturn(somePrivateUrl());
-
-		//Test that 401 Unauthorized is returned (Bad Private token)
-		mockMvc.perform(get("/{id}", "someKey")
-				.param("privateToken","incorrectToken"))
-				.andDo(print())
-				.andExpect(status().isForbidden());
-	}
-
-	@Test
-	/*
-	Test that REDIRECT over a PRIVATE link gives error 401 if key exists
-	and Private Token IS NOT SUPPLIED.
-	 */
-	public void thatRedirectToPrivateReturnsTemporaryRedirectIfKeyExistsAndPrivateTokenNotSupplied()
-			throws Exception {
-
-		//Mock URLrepository response to a private URL.
-		when(shortURLRepository.findByHash("someKey")).thenReturn(somePrivateUrl());
-
-		//Test that 401 Unauthorized is returned (Bad Private token)
-		mockMvc.perform(get("/{id}", "someKey"))
-				.andDo(print())
-				.andExpect(status().isForbidden());
-	}
-
-	@Test
-	/*
-	Test that REDIRECT over an ID that does NOT EXIST gives error 404.
-	 */
-	public void thatRedirectToReturnsNotFoundIdIfKeyDoesNotExist()
-			throws Exception {
-
-		//Mock URLRepository to return null -> Not found
-		when(shortURLRepository.findByHash("someKey")).thenReturn(null);
-
-		mockMvc.perform(get("/{id}", "someKey")).andDo(print())
-				.andExpect(status().isNotFound());
-	}
 
 	@Test
 	/*
